% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_port_be_opened.R
\name{can_port_be_opened}
\alias{can_port_be_opened}
\title{Check whether a TCP port is free or not}
\usage{
can_port_be_opened(port)
}
\arguments{
\item{port}{(integer) A TCP port in [1, 65535].}
}
\value{
\code{can_port_be_opened(port)} returns a logical indicating whether the port
can be opened or not, or cannot be queried.  If the port can be opened,
then \code{TRUE} is returned, if cannot be opened then \code{FALSE} is returned,
which may happen if the port is used by another process.
If port querying is not supported, as in R (< 4.0.0),  then \code{NA} is
returned.
}
\description{
Check whether a TCP port is free or not
}
\keyword{internal}
