% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port4me.R
\name{port4me}
\alias{port4me}
\title{Gets a personalized TCP port that can be opened}
\usage{
port4me(
  user = port4me_user(),
  tool = port4me_tool(),
  prepend = port4me_prepend(),
  include = port4me_include(),
  exclude = port4me_exclude(),
  skip = port4me_skip(),
  list = NULL,
  test = NULL,
  max_tries = 65535L,
  must_work = TRUE
)
}
\arguments{
\item{user}{(optional) The name of the user.
Defaults to \code{Sys.info()[["user"]]}.}

\item{tool}{(optional) The name of the software tool for which a port
should be generated.}

\item{prepend}{(optional) An integer vector of ports to always consider.}

\item{include}{(optional) An integer vector of possible ports to return.
Defaults to \verb{1024::65535}.}

\item{exclude}{(optional) An integer vector of ports to exclude.}

\item{skip}{(optional) Number of non-excluded ports to skip.}

\item{list}{(optional) Number of ports to list.}

\item{test}{(optional) A port to check whether it can be opened or not.}

\item{max_tries}{Maximum number of ports checked, before giving up.}

\item{must_work}{If TRUE, then an error is produced if no port could
be found.  If FALSE, then \code{-1} is returned.}
}
\value{
An port or a vector of ports.
If \code{test} is given, then TRUE is if the port can be opened, otherwise FALSE.
}
\description{
Gets a personalized TCP port that can be opened
}
\examples{
port <- port4me()
print(port)

port <- port4me(tool = "rstudio")
print(port)

ports <- port4me(tool = "rstudio", list = 5L)
print(ports)

avail <- port4me(test = 4321)
print(avail)


}
